Move getAllEmojis function to slack api file
